int was replaced by size_t

+++ tasks/romanova_v_min_by_matrix_rows_processes/seq/src/ops_seq.cpp
@@ -31,9 +31,9 @@ bool RomanovaVMinByMatrixRowsSEQ::PreProcessingImpl() {
 
 bool RomanovaVMinByMatrixRowsSEQ::RunImpl() {
   int min_val;
-  for(int i = 0; i < n_; i++){
+  for(size_t i = 0; i < n_; i++){
     min_val = in_data_[i][0];
-    for(int j = 1; j < m_; j++){
+    for(size_t j = 1; j < m_; j++){
       if(in_data_[i][j] < min_val) min_val = in_data_[i][j];
     }
     res_[i] = min_val;
